Add ability to use Passenger Enterprise
Passenger Enterprise uses a different repository which requires
authentication, and a license file that needs to be installed on the
server.

This commit adds two new attributes, `enterprise_token` and
`enterprise_license` which need to be set to install Passenger
Enterprise.

Versions 4 and 5 of Passenger are still supported for both OSS and
Enterprise.